replace build script with rustfmt and meta tests
