+++ .travis.yml
@@ -7,7 +7,11 @@ env:
     - RUST_BACKTRACE=1
     - secure: E33II/ctVUNigc5wuh153OX/7RUje43G3I9i5+PibEiFPCpKIJpHZA9LfKHABqOBVPgBLdWkretWgKn4Whfb9e2NMC8Nzn8Pch68b7yPZ2Ee0zLWr2M2gJ95PwTn3ZyFaMuZOAa+TgEU9buVIXTdgM+FHK+jXwU0qqNL2XCF5YM=
 
+before_script:
+  - rustup component add rustfmt-preview
+
 script:
+  - cargo fmt --all -- --write-mode=diff
   - cargo build
   - cargo test --all
