+++ Cargo.toml
@@ -1,7 +1,6 @@
 [package]
 name = "rust-rosetta"
 version = "0.0.1"
-build = "build.rs"
 authors = [
     "Andrew Hobden, https://github.com/Hoverbear",
     "Adolfo Ochagavía, https://github.com/aochagavia",
@@ -84,11 +83,9 @@ serde = "1"
 serde_json = "1"
 term = "0.4"
 
-[build-dependencies]
-failure = "0.1"
-meta = { path = "meta" }
+[dev-dependencies]
+lazy_static = "1"
 toml = "0.4"
-unicode-segmentation = "1"
 
 [workspace]
 members = [
